Add Content to Film State, arrange PopOut format
I added most of the information for the first two lists to the film
state, which is passed down to the popout. Arranged this information.
Downloaded a lot of photos to use in the popout. Also created an Overlay
so that the user cannot scroll underneath the popout, but can click
popout to exit as Well