A007: Add related section for film FilmPopOut
Added related section. Underneath film descriptions in the FilmPopOut
is a list of three related movies. Underneath director descriptions is
a list of four recommended movies. Each recommendation/related has an
accompanying image, so a lot of images added to the public folder.
Minor css changes to accommodate this related section, which largely
mimics the structure of the book section