A010: Fix header problem for Code / Blog pages
Those pages were altering the header top margin on toggleblog/code func.
Upon investigation discovered that it was actually a footer problem.
The css change was making visually obvious a problem between the default
margin of the ul and the smaller fixed height of the footer.
Eliminated footer ul margins and also increased the height of the
footer. Problem solved.