A018: go around app fixing console errors
Fixed mainly key-prop errors. Mostly added index to as a parameter to
map methods, and used that as keys, where there was no UUID (because
they were children, mostly.
There was also a prominent error associated with the CSSTransition
usage. It didn't have an affect and according to the internet won't
have an effect on production build (detected in dev because of a
feature called React.StrictMode). Tried to fix it unsuccessfully using
a technique involving changing the children within the transitions to
functional components (rather than react class components). Left the
changed components (BlogItem and FilmPopOut) as functional components,
as it does no harm.