A019: revise code and blog onclick open close
Actually wasn't a problem with the onclick. The <a> tag
was for the purpose of jumping to the top of the blog/project when
it was clicked. Error was because <a> cannot be embedded within <a>,
and one of the projects was linking to some external resources.
Solved by moving the outer <a> to the closed box condition (previously
was covering both conditions outside ternery operator)