A020(b): change fonts
Changed font of film popout for films, not directors.
Also changed some of the line-heights and letter-spacing.
Changed some for the recommendations as well.
Involved creating a new fpo-director class.

Also changed the z-index of the overlay and footer and fpo, because
noticed the footer was showing above the fpo
now it's footer - 10, overlay - 11, fpo - 12