A022: change react styling to css classes
Refactored almost all react styling to css classes
Involved creating a lot of new css classes and some new conditional
class adding in some instances.
Left a couple of instances of react styling where it would have been
irrational to change it, but changed almost all instances
This was the first instance of a styling refactor on this project.
No doubt there will be more in the future