A025: resp. dsn. cool stuff pockets
A big commit. A lot of CSS PORTRAIT media queries set targeting
different sizes for the cool stuff pockets. Not all work on this done
but most (landscape still required).
Most of the work involved changing size of certain photos and
alignment, as well as the size of the overall pocket.