A036/7: mob header update: colors and titleclick
On mobile, AppLinks/Name-Title now remain colored if open
This was accomplished using ComponentDidMount/WillUnmount, and useEffect
in functional components, with a series of useStates in the App
component.
It involved changing the routing slightly in the App component

Also, problem fixed with clicking the Name-Title not closing the
mobile nav menu.

While trying to separate these commits, I made a mistake with Git and
deleted the header file. Was able to recover it, but it is now
being committed as a totally new file, I believe