A039d: landscape-coolstuff resp. dsn
Huge job. Added CSS Media queries and classes to account for different
landscape configurations of the Cool Stuff page.
No doubt there are some bugs, but at least everything is mostly fixed
now, including all font sizes and image sizes.
Also includes a fix for a mistake I made with the earlier Portrait
version of this commit - which was display:none of the cover container.
The right cover is crucial for hiding the right pocket.